src/requirements.txt
fastapi>=0.110
uvicorn>=0.29
scikit-learn>=1.4
numpy>=1.26
joblib>=1.3
psycopg2-binary>=2.9
python-decouple>=3.8
apscheduler>=3.10
